wip fixing model errors
