+++ api/scorer_weighted/views.py
@@ -3,7 +3,7 @@ import logging
 # from django_filters.rest_framework import DjangoFilterBackend
 # from rest_framework import viewsets
 
-from scorer_weighted.models import Score, WeightedScorer
+from scorer_weighted.models import WeightedScorer
 # from scorer_weighted.serializers import ScoreSerializer, WeightedScorerSerializer
 
 log = logging.getLogger(__name__)
